+++ pkg/gatewaysecret/handler.go
@@ -6,6 +6,9 @@ import (
 	"fmt"
 	"time"
 
+	certmanagerv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
+	"github.com/cert-manager/cert-manager/pkg/client/clientset/versioned"
+	certmanagerclientv1 "github.com/cert-manager/cert-manager/pkg/client/clientset/versioned/typed/certmanager/v1"
 	"github.com/go-logr/logr"
 	apicorev1 "k8s.io/api/core/v1"
 	apimetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -18,22 +21,20 @@ import (
 
 const (
 	LastModifiedAtAnnotation = "lastModifiedAt"
-
-	// TODO move to config?
-	kcpCACertName = "klm-watcher-serving"
+	kcpCACertName            = "klm-watcher-serving"
 )
 
 var errCouldNotGetLastModifiedAt = errors.New("getting lastModifiedAt time failed")
 
 type Handler struct {
-	certManagerClient *CertManagerClient
+	certificateClient certmanagerclientv1.CertificateInterface
 	kcpSecretClient   v1.SecretInterface
 	log               logr.Logger
 }
 
 func NewGatewaySecretHandler(config *rest.Config, log logr.Logger) *Handler {
 	return &Handler{
-		certManagerClient: NewCertManagerClient(config),
+		certificateClient: versioned.NewForConfigOrDie(config).CertmanagerV1().Certificates(istioNamespace),
 		kcpSecretClient:   kubernetes.NewForConfigOrDie(config).CoreV1().Secrets(istioNamespace),
 		log:               log,
 	}
@@ -48,7 +49,7 @@ func (h *Handler) ManageGatewaySecret(ctx context.Context, rootSecret *apicorev1
 		return err
 	}
 
-	caCert, err := h.certManagerClient.GetRootCACertificate(ctx)
+	caCert, err := h.GetRootCACertificate(ctx)
 	if err != nil {
 		return err
 	}
@@ -68,6 +69,14 @@ func (h *Handler) findGatewaySecret(ctx context.Context) (*apicorev1.Secret, err
 	return secret, nil
 }
 
+func (h *Handler) GetRootCACertificate(ctx context.Context) (*certmanagerv1.Certificate, error) {
+	caCert, err := h.certificateClient.Get(ctx, kcpCACertName, apimetav1.GetOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get CA certificate: %w", err)
+	}
+	return caCert, nil
+}
+
 func (h *Handler) Create(ctx context.Context, secret *apicorev1.Secret) error {
 	setLastModifiedToNow(secret)
 	if _, err := h.kcpSecretClient.Create(ctx, secret, apimetav1.CreateOptions{}); err != nil {
@@ -82,6 +91,15 @@ func (h *Handler) Update(ctx context.Context, secret *apicorev1.Secret) error {
 		return fmt.Errorf("failed to update secret %s: %w", secret.Name, err)
 	}
 	return nil
+}
+
+func RequiresUpdate(gwSecret *apicorev1.Secret, caCert *certmanagerv1.Certificate) bool {
+	if gwSecretLastModifiedAt, err := ParseLastModifiedTime(gwSecret); err == nil {
+		if caCert.Status.NotBefore != nil && gwSecretLastModifiedAt.After(caCert.Status.NotBefore.Time) {
+			return false
+		}
+	}
+	return true
 }
 
 func CopySecretData(from, to *apicorev1.Secret) {
